virtualchannels/requirements.txt
pyln-client
pyln-proto >= 0.8.2
pyln-testing
